Fix slow network failures, issue #8
IT now downloads upto 3 files at a time to support the slow connection.  and maintains the number of downloads plus items not played on the list so that it will not run out of songs.  Has been checked with Cached and Saved.  This change does not affect the stream setting.  I have not figured out how to make streaming work with slow connections since downloads seem to be 50% of the stream rate from the Pandora server.